v5: fix implicit conversion changes signedness build error
The Android toolchain compiler reports the following error:

FAILED:
out/soong/.intermediates/hardware/google/apf/v5/libapf_v5/android_x86_64_silvermont_sdk_static/obj/hardware/google/apf/v5/apf_interpreter.o
hardware/google/apf/v5/apf_interpreter.c:771:85: error: implicit conversion changes signedness: 'u32' (aka 'unsigned int') to 's32' (aka
'int') [-Werror,-Wsign-conversion]
int dscp = calculate_checksum_and_return_dscp(allocated_buffer, pkt_len);

Test: TH

+++ v5/apf_interpreter.c
@@ -801,7 +801,10 @@ int apf_run(void* ctx, u8* const program, const u32 program_len,
                             0 /* dscp */);
                         return PASS_PACKET;
                     }
-                    int dscp = calculate_checksum_and_return_dscp(allocated_buffer, pkt_len);
+                    // allocated_buffer_len cannot be large because we'd run out of RAM,
+                    // so the above unsigned comparison effectively guarantees casting pkt_len
+                    // to a signed value does not result in it going negative.
+                    int dscp = calculate_checksum_and_return_dscp(allocated_buffer, (s32)pkt_len);
                     apf_transmit_buffer(ctx, allocated_buffer, pkt_len, dscp);
                     allocated_buffer = NULL;
                     break;
